logging-service/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
uvicorn==0.17.6
hazelcast-python-client==5.1.0